refactor trace file backend import
